Switched to .hpp and paging clean up
