+++ kernel/gdt/gdt.cpp
@@ -1,5 +1,4 @@
-
-#include "gdt.h"
+#include "gdt.hpp"
  
 void create_descriptor()
 {
