+++ kernel/memory/memory.cpp
@@ -1,4 +1,4 @@
-#include "memory.h"
+#include "memory.hpp"
 
 void memset(void* start, uint8_t value, uint64_t num){
     for (uint64_t i = 0; i < num; i++){
