+++ kernel/paging/global_addresses.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#define KEY_INFO_ADDRESS 0x1600
+#define RTC_DATETIME_AREA 0x1610
+#define IRQ0_SLEEP_TIMER_TICKS_AREA 0x1700
+#define CURRENT_PAGE_DIR_ADDRESS 0x1800
+#define PHYS_MEM_MAX_BLOCKS 0x1804
+#define PHYS_MEM_USED_BLOCKS 0x1808
+#define MEMMAP_AREA 0x30000
+#define KERNEL_ADDRESS 0x100000 // 1MB
