--- kernel/test.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "limine.h"
-#include "e9print.h"
-
-void test();
