+++ kernel/test.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "e9print.hpp"
+
+void test();
